requirements.txt
torch==2.0.1
opencv-python
apache-tvm
onnx2torch
ultralytics